requirements.txt
lxml==4.3.4
Pillow==6.0.0
python-pptx==0.6.18
XlsxWriter==1.1.8
